Refactored directory to combine files and rename files so that they make more sense.
