+++ src/hex/grid.rs
@@ -2,7 +2,7 @@
 
 use std::collections::HashMap;
 
-use coordinate::Axial;
+use axial::Axial;
 
 use super::*;
 
@@ -49,7 +49,7 @@ impl<TileType> HexGrid<TileType> {
     /// ```
     /// let my_object_pos = (100.0, 432.0);
     /// /// ...
-    /// use gridava::hex::hex_grid::*;
+    /// use gridava::hex::grid::*;
     ///
     /// let my_grid = HexGrid::<i32>::default();
     /// let nearest_tile = my_grid.world_to_hex(my_object_pos);
@@ -93,8 +93,8 @@ impl<TileType> HexGrid<TileType> {
     /// # Example
     /// ```
     /// /// ...
-    /// use gridava::hex::hex_grid::*;
-    /// use gridava::hex::coordinate::*;
+    /// use gridava::hex::grid::*;
+    /// use gridava::hex::axial::*;
     ///
     /// let my_grid = HexGrid::<i32>::default();
     /// let world_pos = my_grid.hex_to_world(axial!(12, 33));
