+++ src/hex/mod.rs
@@ -2,8 +2,6 @@
 //!
 //! TODO: Examples.
 //!
-pub mod algorithms;
-pub mod coordinate;
-pub mod hex_grid;
+pub mod axial;
+pub mod grid;
 pub mod shape;
-pub mod shape_constructors;
